Split code :art: for later compatibility with multiple operating systems
